Struct tst::TSTSet

pub struct TSTSet {
    // some fields omitted
}

A set based on a TSTMap.

Methods

impl TSTSet

fn new() -> TSTSet

Makes a new empty TSTSet.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();

fn len(&self) -> usize

Returns the number of elements in the set.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();
assert_eq!(s.len(), 0);
s.insert("xxx");
assert_eq!(s.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();
assert!(s.is_empty());
s.insert("yyyx");
assert!(!s.is_empty());

fn clear(&mut self)

Clears the set, removing all values.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();
s.insert("abc");
s.insert("abd");
s.clear();

assert!(s.is_empty());
assert!(!s.contains("abc"));

fn contains(&self, key: &str) -> bool

Returns true if the set contains a value.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();
s.insert("abc");
assert!(!s.contains("ab"));
assert!(s.contains("abc"));

fn insert(&mut self, key: &str) -> bool

Adds a value to the set.

If the set did not have a value present, true is returned.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();

assert!(s.insert("abcd"));
assert!(!s.insert("abcd"));
assert_eq!(s.len(), 1);

fn remove(&mut self, key: &str) -> bool

Removes a value from the set. Returns true if the value was present in the set.

Examples

use tst::TSTSet;

let mut s: TSTSet = TSTSet::new();

s.insert("acde");
assert!(s.remove("acde"));
assert!(!s.remove("acde"));